auth/requirements.txt
blinker==1.9.0
click==8.1.8
Flask==3.1.0
Flask-MySQLdb==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
mysqlclient==2.2.7
PyJWT==2.10.1
PyMySQL==1.1.1
Werkzeug==3.1.3
